Change float to double
Floats:🤮
Doubles: 😄

+++ TeamCode/src/main/java/t10/motion/hardware/PositionalMotor.java
@@ -12,13 +12,13 @@ import t10.utils.PIDController;
 public class PositionalMotor {
     private final DcMotorEx motor;
     private final PIDController pid;
-    private final float offsetTicks;
-    private final float ticksPerRotation;
-    private final float minBoundTicks;
-    private final float maxBoundTicks;
-    private final float speed;
-    private float currentPositionTicks = 0;
-    private float targetTicks;
+    private final double offsetTicks;
+    private final double ticksPerRotation;
+    private final double minBoundTicks;
+    private final double maxBoundTicks;
+    private final double speed;
+    private double currentPositionTicks = 0;
+    private double targetTicks;
 
     /**
      * Creates a PositionalMotor
@@ -29,7 +29,7 @@ public class PositionalMotor {
      * @param initialRotation the initial position, in rotations, of the motor
      * @param speed the speed for the motor to rotate at (can be negative for inverted rotation)
      */
-    public PositionalMotor(DcMotorEx motor, float ticksPerRotation, float minBoundRotation, float maxBoundRotation, float initialRotation, float speed) {
+    public PositionalMotor(DcMotorEx motor, double ticksPerRotation, double minBoundRotation, double maxBoundRotation, double initialRotation, double speed) {
         if (minBoundRotation >= maxBoundRotation) {
             throw new InvalidParameterException("minBoundRotation must be less than maxBoundRotation");
         }
@@ -46,15 +46,15 @@ public class PositionalMotor {
         this.pid = new PIDController(0.1, 0, 0);
     }
 
-    private float rotationToTicks(float rotation) {
+    private double rotationToTicks(double rotation) {
         return rotation * ticksPerRotation;
     }
 
-    private float ticksToRotation(float ticks) {
+    private double ticksToRotation(double ticks) {
         return ticks / ticksPerRotation;
     }
 
-    public void setPosition(float rotation) {
+    public void setPosition(double rotation) {
         if (rotation < ticksToRotation(minBoundTicks) || rotation > ticksToRotation(maxBoundTicks)) {
             throw new InvalidParameterException("Position is not within bounds [" + ticksToRotation(minBoundTicks) + "," + ticksToRotation(maxBoundTicks) + "]");
         }
@@ -63,14 +63,14 @@ public class PositionalMotor {
     }
 
     public void update() {
-        float ticks = targetTicks - currentPositionTicks;
+        double ticks = targetTicks - currentPositionTicks;
 
         motor.setVelocity(ticks * speed);
 
         currentPositionTicks = motor.getCurrentPosition() + offsetTicks;
     }
 
-    public float getPosition() {
+    public double getPosition() {
         return ticksToRotation(currentPositionTicks);
     }
 }
